Refactor extensionHost code and protocol detection code for readability
