+++ src/node/extension/extension.ts
@@ -4,15 +4,12 @@
 
 'use strict';
 
-import * as net from 'net';
 import * as vscode from 'vscode';
-import { spawn, spawnSync, exec } from 'child_process';
+import { spawn, exec } from 'child_process';
 import { basename, join, isAbsolute, dirname } from 'path';
-import * as nls from 'vscode-nls';
 import * as fs from 'fs';
-
-const localize = nls.config(process.env.VSCODE_NLS_CONFIG)();
-
+import { log, localize } from './utilities';
+import { determineDebugType } from './protocolDetection';
 
 export function activate(context: vscode.ExtensionContext) {
 
@@ -367,9 +364,6 @@ function guessProgramFromPackage(jsonObject: any): string | undefined {
 
 //---- extension.node-debug.startSession
 
-// For launch, use inspector protocol starting with v8 because it's stable after that version.
-const InspectorMinNodeVersionLaunch = 80000;
-
 /**
  * The result type of the startSession command.
  */
@@ -381,36 +375,7 @@ class StartSessionResult {
 function startSession(config: any): StartSessionResult {
 
 	if (Object.keys(config).length === 0) { // an empty config represents a missing launch.json
-
-		config.type = 'node';
-		config.name = 'Launch';
-		config.request = 'launch';
-
-		if (vscode.workspace.rootPath) {
-
-			// folder case: try to find more launch info in package.json
-			const pkg = loadPackage(vscode.workspace.rootPath);
-			if (pkg) {
-				if (pkg.name === 'mern-starter') {
-					configureMern(config);
-				} else {
-					config.program = guessProgramFromPackage(pkg);
-				}
-			}
-		}
-
-		if (!config.program) {
-
-			// 'no folder' case (or no program found)
-			const editor = vscode.window.activeTextEditor;
-			if (editor && editor.document.languageId === 'javascript') {
-				config.program = editor.document.fileName;
-			} else {
-				return {
-					status: 'initialConfiguration'	// let VS Code create an initial configuration
-				};
-			}
-		}
+		config = getFreshLaunchConfig();
 	}
 
 	// make sure that 'launch' configs have a 'cwd' attribute set
@@ -423,60 +388,9 @@ function startSession(config: any): StartSessionResult {
 		}
 	}
 
-	// determine what protocol to use
-
-	let fixConfig = Promise.resolve<any>();
-
-	switch (config.protocol) {
-		case 'legacy':
-			config.type = 'node';
-			break;
-		case 'inspector':
-			config.type = 'node2';
-			break;
-		case 'auto':
-		default:
-			config.type = 'node';
-
-			switch (config.request) {
-
-				case 'attach':
-					fixConfig = getProtocolForAttach(config).then(protocol => {
-						if (protocol === 'inspector') {
-							config.type = 'node2';
-						}
-					});
-					break;
-
-				case 'launch':
-					if (config.runtimeExecutable) {
-						config.type = 'node2';
-						log(localize('protocol.switch.runtime.set', "Debugging with inspector protocol because a runtime executable is set."));
-					} else {
-						// only determine version if no runtimeExecutable is set (and 'node' on PATH is used)
-						const result = spawnSync('node', [ '--version' ]);
-						const semVerString = result.stdout.toString();
-						if (semVerString) {
-							if (semVerStringToInt(semVerString) >= InspectorMinNodeVersionLaunch) {
-								config.type = 'node2';
-								log(localize('protocol.switch.inspector.version', "Debugging with inspector protocol because Node.js {0} was detected.", semVerString.trim()));
-							} else {
-								log(localize('protocol.switch.legacy.version', "Debugging with legacy protocol because Node.js {0} was detected.", semVerString.trim()));
-							}
-						} else {
-							log(localize('protocol.switch.unknown.version', "Debugging with legacy protocol because Node.js version could not be determined."));
-						}
-					}
-					break;
-
-				default:
-					// should not happen
-					break;
-			}
-			break;
-	}
-
-	fixConfig.then(() => {
+	// determine which protocol to use
+	determineDebugType(config).then(debugType => {
+		config.type = debugType;
 		vscode.commands.executeCommand('vscode.startDebug', config);
 	});
 
@@ -485,83 +399,40 @@ function startSession(config: any): StartSessionResult {
 	};
 }
 
-function log(message: string) {
-	vscode.commands.executeCommand('debug.logToDebugConsole', message + '\n');
-}
-
-/**
- * Detect which debug protocol is being used for a running node process.
- */
-function getProtocolForAttach(config: any): Promise<string|undefined> {
-	const address = config.address || '127.0.0.1';
-	const port = config.port;
-
-	if (config.processId) {
-		// this is only supported for legacy protocol
-		log(localize('protocol.switch.attach.process', "Debugging with legacy protocol because attaching to a process by ID is only supported for legacy protocol."));
-		return Promise.resolve('legacy');
-	}
-
-	const socket = new net.Socket();
-	const cleanup = () => {
-		try {
-			socket.write(`"Content-Length: 50\r\n\r\n{"command":"disconnect","type":"request","seq":2}"`);
-			socket.end();
-		} catch (e) {
-			// ignore failure
-		}
+function getFreshLaunchConfig(): any {
+	const config: any = {
+		type: 'node',
+		name: 'Launch',
+		request: 'launch'
 	};
 
-	return new Promise<{reason: string, protocol: string}>((resolve, reject) => {
-		socket.once('data', data => {
-			let reason: string;
-			let protocol: string;
-			const dataStr = data.toString();
-			if (dataStr.indexOf('WebSockets request was expected') >= 0) {
-				reason = localize('protocol.switch.inspector.detected', "Debugging with inspector protocol because it was detected.");
-				protocol = 'inspector';
+	if (vscode.workspace.rootPath) {
+
+		// folder case: try to find more launch info in package.json
+		const pkg = loadPackage(vscode.workspace.rootPath);
+		if (pkg) {
+			if (pkg.name === 'mern-starter') {
+				configureMern(config);
 			} else {
-				reason = localize('protocol.switch.legacy.detected', "Debugging with legacy protocol because it was detected.");
-				protocol = 'legacy';
+				config.program = guessProgramFromPackage(pkg);
 			}
-
-			resolve({ reason, protocol });
-		});
-
-		socket.once('error', err => {
-			reject(err);
-		});
-
-		socket.connect(port, address);
-		socket.on('connect', () => {
-			// Send a safe request to trigger a response from the inspector protocol
-			socket.write(`Content-Length: 102\r\n\r\n{"command":"evaluate","arguments":{"expression":"process.pid","global":true},"type":"request","seq":1}`);
-		});
-
-		setTimeout(() => {
-			// No data or error received? Bail and let the debug adapter handle it.
-			reject(new Error('timeout'));
-		}, 2000);
-	}).catch(err => {
-		return {
-			reason: localize('protocol.switch.unknown.error', "Debugging with legacy protocol because Node.js version could not be determined ({0})", err.toString()),
-			protocol: 'legacy'
-		};
-	}).then(result => {
-		cleanup();
-		log(result.reason);
-
-		return result.protocol;
-	});
-}
-
-/**
- * convert the 3 parts of a semVer string into a single number
- */
-function semVerStringToInt(vString: string): number {
-	const match = vString.match(/v(\d+)\.(\d+)\.(\d+)/);
-	if (match && match.length === 4) {
-		return (parseInt(match[1])*100 + parseInt(match[2]))*100 + parseInt(match[3]);
+		}
 	}
-	return -1;
+
+	if (!config.program) {
+
+		// 'no folder' case (or no program found)
+		const editor = vscode.window.activeTextEditor;
+		if (editor && editor.document.languageId === 'javascript') {
+			config.program = editor.document.fileName;
+		} else {
+			return {
+				status: 'initialConfiguration'	// let VS Code create an initial configuration
+			};
+		}
+	}
+
+	return config;
 }
+
+
